refactor(web): extract session-toc from SessionDetail
Move SessionToc, SessionTocFilterPanel, toggleTocFilter (+ private
TocCheckbox/TOC_META/getTocToolIds) into session-detail/session-toc.tsx.
SessionToc renders the already-extracted FileChangeTracker. Pure move +
import rewire. Second sub-domain of CS-13.
